Update README.md
The  Nuclear Stress Tester is a powerful and highly customizable HTTP load testing tool for developers and DevOps professionals. This script leverages the power of **Vegeta**, a versatile HTTP load testing library, and combines it with a user-friendly Bash interface to provide dynamic stress testing capabilities.

With an intuitive menu system, colorful interface, and robust reporting features, this tool is designed to help you analyze the performance and scalability of your web applications or APIs.